# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-inrange",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — decide whether `value` is within `[lo, hi]`.",
        "description": "This service is a thin orchestrator over `srvcs-between`: it does no\ncomparison of its own and delegates entirely, forwarding\n`{\"value\", \"lo\", \"hi\"}` to `srvcs-between` and returning its boolean\n`result`. Invalid operands are rejected by `srvcs-between` and the resulting\n`422` is forwarded unchanged.",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "an operand is invalid (forwarded from srvcs-between)"
          },
          "500": {
            "description": "a dependency returned a malformed result"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EvalRequest": {
        "type": "object",
        "required": [
          "value",
          "lo",
          "hi"
        ],
        "properties": {
          "hi": {
            "type": "object"
          },
          "lo": {
            "type": "object"
          },
          "value": {
            "type": "object"
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      },
      "ResultResponse": {
        "type": "object",
        "required": [
          "value",
          "lo",
          "hi",
          "result"
        ],
        "properties": {
          "hi": {
            "type": "object"
          },
          "lo": {
            "type": "object"
          },
          "result": {
            "type": "boolean",
            "description": "`true` iff `lo <= value <= hi`."
          },
          "value": {
            "type": "object"
          }
        }
      }
    }
  }
}